% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompdist.R
\name{gomp}
\alias{gomp}
\alias{dgomp}
\alias{pgomp}
\alias{qgomp}
\alias{rgomp}
\title{The Gompertz Distribution}
\usage{
dgomp(x, eta, b = 1)

pgomp(q, eta, b = 1)

qgomp(p, eta, b = 1)

rgomp(n, eta, b = 1)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{eta}{shape parameter}

\item{b}{scale parameter}

\item{p}{vector of probabilities}

\item{n}{number of observations}
}
\value{
\code{dgomp} gives the density function, \code{pgomp} gives the distribution
function, \code{qgomp} gives the quantile function, and \code{rgomp} generates random values
}
\description{
Density, distribution, and quantile functions along with random number
generation for the Gompertz distribution.
}
\examples{
x = seq(0, 5, by = 0.1)
dgomp(x, eta = 0.2)
p <- 0:10 * 1 / 10
pgomp(qgomp(p, eta = 2), eta = 2)
rgomp(10, eta = 0.2)
}
