% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gomp_fit_wrapper.R
\name{gomp_fit}
\alias{gomp_fit}
\title{Fit Data to the Gompertz Distribution}
\usage{
gomp_fit(
  x,
  y,
  parstart = NULL,
  eps = 1e-04,
  lambda = 1,
  lamup = 1.1,
  lamdown = 1.5,
  maxIter = 100
)
}
\arguments{
\item{x}{vector of ages}

\item{y}{vector of number of deaths at age x}

\item{parstart}{vector of starting values for parameters \code{eta} and \code{b} respectively}

\item{eps}{stopping criterion}

\item{lambda}{damping parameter}

\item{lamup}{multiplied with lambda to decrease step size if the previous step is not accepted}

\item{lamdown}{lambda is divided by this to increase step size if the previous step is accepted}

\item{maxIter}{the maximum number of times the algorithm will iterate}
}
\value{
List containing the vector of estimates for \code{eta} and \code{b} and the MSE at convergence.
}
\description{
Fit mortality data to the Gompertz distribution using
nonlinear least squares. The algorithm used here is the Levenberg-Marquardt
algorithm. See \url{https://doi.org/10.1103/PhysRevE.83.036701} for more info.
}
\examples{
x = seq(0, 5, by = 0.2)
y = abs(dgomp(x, eta = 0.1, b = 1) + rnorm(length(x), mean = 0, sd = 0.01))
gomp_fit(x, y)
}
