% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompmdist.R
\name{gompm}
\alias{gompm}
\alias{dgompm}
\alias{pgompm}
\alias{qgompm}
\alias{rgompm}
\title{The Gompertz-Makeham Distribution}
\usage{
dgompm(x, alpha, beta, lambda)

pgompm(q, alpha, beta, lambda)

qgompm(p, alpha, beta, lambda)

rgompm(n, alpha, beta, lambda)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{alpha, beta, lambda}{parameters}

\item{p}{vector of probabilities}

\item{n}{number of observations}
}
\value{
\code{dgompm} gives the density function, \code{pgompm} gives the distribution
function, \code{qgompm} gives the quantile function, and \code{rgompm} generates random values
}
\description{
Density, distribution, and quantile functions along with random number
generation for the Gompertz-Makeham distribution.
}
\examples{
x = seq(0, 5, by = 0.1)
dgompm(x, 1, 1, 1)
p <- 0:10 * 1 / 10
pgompm(qgompm(p, .5, .6, .7), .5, .6, .7)
rgompm(10, 1, 2, 3)
}
